Mayor factoring & added GPL headers

+++ include/gpu/gpu_structures.h
@@ -1,9 +1,27 @@
 /*
- * PROJECT: GEMMapper
- * FILE: gpu_structures.h
- * DATE: 06/06/2012
- * AUTHOR(S): Alejandro Chacon <alejandro.chacon@uab.es>
- *            Santiago Marco-Sola <santiagomsola@gmail.com>
+ *  GEM-Mapper v3 (GEM3)
+ *  Copyright (c) 2011-2017 by Santiago Marco-Sola  <santiagomsola@gmail.com>
+ *  Copyright (c) 2011-2017 by Alejandro Chacon <alejandro.chacon@uab.es>
+ *
+ *  This file is part of GEM-Mapper v3 (GEM3).
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * PROJECT: GEM-Mapper v3 (GEM3)
+ * AUTHOR(S): Santiago Marco-Sola <santiagomsola@gmail.com>
+ * DESCRIPTION:
+ *   GPU-adaptor module provides functions to write a GEM3-GPU index
  */
 
 #ifndef GPU_STRUCTURES_H_
@@ -12,7 +30,7 @@
 #include "utils/essentials.h"
 #include "fm_index/bwt.h"
 #include "fm_index/rank_mtable.h"
-#include "data_structures/dna_text.h"
+#include "text/dna_text.h"
 
 /*
  * GPU Structures Write
@@ -24,6 +42,6 @@ void gpu_structures_write(
     bwt_builder_t* const bwt_builder,
     rank_mtable_t* const rank_mtable,
     uint64_t* const sa_gem,
-    const uint32_t sa_sampling);
+    const uint64_t sa_sampling);
 
 #endif /* GPU_STRUCTURES_H_ */
